Tidy up combat dependencies
