+++ src/models/v4/combat-state/current-combat-state.ts
@@ -1,7 +1,10 @@
 import type { Team } from '../../team';
-import type { ActiveBuffCollection } from '../buff/active-buff-collection';
+import type { AttackAbility } from '../attack/attack-ability';
+import { ActiveBuffCollection } from '../buff/active-buff-collection';
+import type { BuffAbility } from '../buff/buff-ability';
 import type { Character } from '../character/character';
 import type { Charge } from '../charge/charge';
+import type { Registry } from '../registry/registry';
 import type { Resource } from '../resource/resource';
 import type { Target } from '../target/target';
 import type { WeaponTracker } from '../weapon-tracker/weapon-tracker';
@@ -11,74 +14,70 @@ export class CurrentCombatState {
   private _value!: CombatState;
 
   public constructor(
-    character: Character,
-    team: Team,
-    weaponTracker: WeaponTracker,
-    target: Target,
-    resources: Resource[],
-    charge: Charge | undefined,
-    activeBuffs: ActiveBuffCollection
+    private readonly character: Character,
+    private readonly team: Team,
+    private readonly weaponTracker: WeaponTracker,
+    private readonly target: Target,
+    private readonly charge: Charge | undefined,
+    private readonly resources: Registry<Resource>,
+    private readonly attacks: Registry<AttackAbility>,
+    private readonly buffs: Registry<BuffAbility>
   ) {
-    this.update(
-      character,
-      team,
-      weaponTracker,
-      target,
-      resources,
-      charge,
-      activeBuffs
-    );
+    this.update();
   }
 
   public get value(): CombatState {
     return this._value;
   }
 
-  public update(
-    character: Character,
-    team: Team,
-    weaponTracker: WeaponTracker,
-    target: Target,
-    resources: Resource[],
-    charge: Charge | undefined,
-    activeBuffs: ActiveBuffCollection
-  ) {
-    const activeWeapon = weaponTracker.getActiveWeapon();
+  public update() {
+    const activeWeapon = this.weaponTracker.getActiveWeapon();
     const activeWeaponState = activeWeapon
       ? { id: activeWeapon.id, damageElement: activeWeapon.damageElement }
       : undefined;
 
-    const previousWeapon = weaponTracker.getPreviousWeapon();
+    const previousWeapon = this.weaponTracker.getPreviousWeapon();
     const previousWeaponState = previousWeapon
       ? { id: previousWeapon.id, damageElement: previousWeapon.damageElement }
       : undefined;
 
+    const activeBuffs = this.getActiveBuffs();
+
     this._value = {
-      team: team,
+      team: this.team,
 
       activeWeapon: activeWeaponState,
       previousWeapon: previousWeaponState,
 
-      resources: resources.map((resource) => ({
+      resources: this.resources.items.map((resource) => ({
         id: resource.id,
         amount: resource.getCumulatedAmount(),
       })),
       charge: {
-        amount: charge?.getCumulatedAmount() ?? 0,
-        hasFullCharge: charge?.hasFullCharge() ?? false,
+        amount: this.charge?.getCumulatedAmount() ?? 0,
+        hasFullCharge: this.charge?.hasFullCharge() ?? false,
       },
 
       activeBuffs,
 
-      elementalAttacks: character.getElementalAttacks(activeBuffs),
-      critRateFlat: character.getCritRateFlat(),
-      critRatePercent: character.getCritRatePercent(activeBuffs),
-      totalCritRatePercent: character.getTotalCritRatePercent(activeBuffs),
-      totalCritDamagePercent: character.getTotalCritDamagePercent(activeBuffs),
-      hp: character.getHp(),
-      elementalResistances: character.getElementalResistances(),
+      elementalAttacks: this.character.getElementalAttacks(activeBuffs),
+      critRateFlat: this.character.getCritRateFlat(),
+      critRatePercent: this.character.getCritRatePercent(activeBuffs),
+      totalCritRatePercent: this.character.getTotalCritRatePercent(activeBuffs),
+      totalCritDamagePercent:
+        this.character.getTotalCritDamagePercent(activeBuffs),
+      hp: this.character.getHp(),
+      elementalResistances: this.character.getElementalResistances(),
 
-      targetResistance: target.resistance,
+      targetResistance: this.target.resistance,
     };
+  }
+
+  private getActiveBuffs() {
+    return new ActiveBuffCollection(
+      this.buffs.items.flatMap(
+        (buffAbility) => buffAbility.getActiveBuffs().items
+      )
+    );
   }
 }
